(Merge tag 'clone-5.7.31-build' into mysql-5.7-cluster-7.5)

+++ sql/protocol_classic.h
@@ -1,7 +1,7 @@
 #ifndef PROTOCOL_CLASSIC_INCLUDED
 #define PROTOCOL_CLASSIC_INCLUDED
 
-/* Copyright (c) 2002, 2018, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2002, 2020, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License, version 2.0,
@@ -202,7 +202,7 @@ public:
   /* Return packet string */
   String *get_packet();
   /* return packet length */
-  uint get_packet_length() { return packet_length; }
+  ulong get_packet_length() { return packet_length; }
   /* Return raw packet buffer */
   uchar *get_raw_packet() { return raw_packet; }
   /* Set read timeout */
